# docs/openapi.yaml
openapi: 3.1.0
info:
  title: RecallAI
  version: 0.1.0
paths:
...
  /api/v1/chat/proposals/{proposal_token}/accept:
    post:
      tags:
      - chat
      summary: Accept Proposal
      description: 'Do the thing the assistant offered to do. No model runs in this path.


        The whole point of a proposal is that the write happens here rather than in a tool a

        model can call: tool results are scraped page text, and a write tool bound to a model

        reading them is one caption away from filing something the person never asked for.


        The path parameter is `proposal_token` and not `token` on purpose. FastAPI resolves

        path-parameter names across the entire dependency tree, and `get_current_user` already

        takes a `token` -- from a cookie, with a default, which a path parameter may not have.

        Same trap as `/spaces/invites/{invite_token}/accept`.'
      operationId: accept_proposal_api_v1_chat_proposals__proposal_token__accept_post
      parameters:
      - name: proposal_token
        in: path
        required: true
        schema:
          type: string
          title: Proposal Token
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/chat/proposals/{proposal_token}/decline:
    post:
      tags:
      - chat
      summary: Decline Proposal
      description: 'Burn the token without doing anything with it.


        Spending it on a decline is the point: it stops a card left open in a tab from being

        tappable later by anyone who reaches that tab.'
      operationId: decline_proposal_api_v1_chat_proposals__proposal_token__decline_post
      parameters:
      - name: proposal_token
        in: path
        required: true
        schema:
          type: string
          title: Proposal Token
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProposalResult'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/chat/ask:
    post:
      tags:
      - chat
      summary: Ask
      description: 'Ask about your own memories.


        Server-sent events: `status` while it searches, `delta` as the answer is written,

        `items` for a listing, and `end`.


        `end` is always sent, including after a provider failure -- a stream that simply

        stops leaves the reader watching a cursor that will never move.'
      operationId: ask_api_v1_chat_ask_post
      parameters:
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AskRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/spaces/invites/{invite_token}/accept:
    post:
      tags:
      - spaces
      summary: Accept Invite
      description: 'Spend an invite link.


        Every rejection is the same 404 -- unknown token, spent token, expired token, deleted

        Space. Distinguishing them tells whoever found the link in a screenshot exactly what

        they are holding.


        Declared **above** the `/{space_id}` routes, like `GET /vault/uploads/limits` is: a

        literal segment that could be read as a parameter has to be matched first. The path

        parameter is `invite_token` rather than `token` because FastAPI resolves path-parameter

        names across the whole dependency tree, and `get_current_user` already takes a `token`

        -- from a cookie, with a default, which a path parameter may not have.'
      operationId: accept_invite_api_v1_spaces_invites__invite_token__accept_post
      parameters:
      - name: invite_token
        in: path
        required: true
        schema:
          type: string
          title: Invite Token
      - name: recall_session
        in: cookie
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Recall Session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpaceRead'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AskRequest:
      properties:
        question:
          type: string
          maxLength: 2000
          minLength: 1
          title: Question
        conversation_id:
          type: string
          maxLength: 64
          minLength: 1
          title: Conversation Id
          default: web
      type: object
      required:
      - question
      title: AskRequest
      description: 'One question. Deliberately the only field.


        No `user_id`, no memory ids, no filters, no model or prompt override: everything that

        decides *which rows are read* comes from the session, and everything that decides how

        they are read is server-side. A request body with nothing in it to tamper with is one

        that cannot be tampered with.'
...
    ContentType:
      type: string
      enum:
      - youtube
      - article
      - pdf
      - document
      - note
      - instagram
      - facebook
      - tiktok
      - linkedin
      - voice
      - image
      title: ContentType
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ProcessingStatus:
      type: string
      enum:
      - pending
      - processing
      - completed
      - failed
      - skipped
      title: ProcessingStatus
    ProposalResult:
      properties:
        status:
          type: string
          title: Status
        message:
          type: string
          title: Message
        item:
          anyOf:
          - $ref: '#/components/schemas/VaultItemRead'
          - type: 'null'
      type: object
      required:
      - status
      - message
      title: ProposalResult
      description: 'What a tapped confirmation did. Deliberately thin.


        It reports an outcome and, when one was created, the card for it -- the same

        `VaultItemRead` every other listing returns, so nothing reaches the browser here that

        a listing would not already show.'
...
    SpaceRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        slug:
          type: string
          title: Slug
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        visibility:
          $ref: '#/components/schemas/Visibility'
        icon:
          anyOf:
          - type: string
          - type: 'null'
          title: Icon
        emoji:
          anyOf:
          - type: string
          - type: 'null'
          title: Emoji
        accent:
          anyOf:
          - type: string
          - type: 'null'
          title: Accent
        pinned:
          type: boolean
          title: Pinned
        ai_overview:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Overview
        ai_topics:
          items:
            type: string
          type: array
          title: Ai Topics
        created_at:
          type: string
          format: date-time
          title: Created At
        role:
          $ref: '#/components/schemas/SpaceRole'
        memory_count:
          type: integer
          title: Memory Count
        member_count:
          type: integer
          title: Member Count
        connection_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Connection Count
      type: object
      required:
      - id
      - name
      - slug
      - description
      - visibility
      - icon
      - emoji
      - accent
      - pinned
      - ai_overview
      - ai_topics
      - created_at
      - role
      - memory_count
      - member_count
      title: SpaceRead
    SpaceRole:
      type: string
      enum:
      - viewer
      - editor
      - owner
      title: SpaceRole
      description: 'What a person may do inside a Space they did not create.


        Ordered least-to-most on purpose so `_RANK` in the service is a comparison rather than

        a table of cases. `owner` is never stored in `space_members` -- it is `spaces.user_id`

        -- but it appears here because the API reports a caller''s effective role and the

        owner''s answer has to be sayable.'
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    VaultItemRead:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        type:
          $ref: '#/components/schemas/ContentType'
        source_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Url
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        summary:
          anyOf:
          - type: string
          - type: 'null'
          title: Summary
        thumbnail_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Thumbnail Url
        ai_tags:
          items:
            type: string
          type: array
          title: Ai Tags
        ai_category:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Category
        ai_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Ai Label
        processing_status:
          $ref: '#/components/schemas/ProcessingStatus'
        processing_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Processing Error
        created_at:
          type: string
          format: date-time
          title: Created At
        file_name:
          anyOf:
          - type: string
          - type: 'null'
          title: File Name
        file_size:
          anyOf:
          - type: integer
          - type: 'null'
          title: File Size
        mime_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Mime Type
      type: object
      required:
      - id
      - type
      - source_url
      - title
      - summary
      - thumbnail_url
      - ai_tags
      - ai_category
      - processing_status
      - created_at
      title: VaultItemRead
...
    Visibility:
      type: string
      enum:
      - private
      - unlisted
      - public
      title: Visibility
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: recall_session
      description: Signed JWT session cookie set by /api/v1/auth/google/callback.